Fix warning if .dockerignore doesn't exist
Happens when changing branches

+++ docker/full/entrypoint.sh
@@ -33,8 +33,10 @@ mkdir -p /myaac-main-folder
 cat > /usr/local/bin/sync-windows-files.sh <<'EOF'
 #!/bin/bash
 while true; do
-  rsync -au --exclude-from="/myaac-main-folder/.dockerignore" "/myaac-main-folder/" "/var/www/html"
-  chown -R www-data:www-data /var/www/html
+  if [ -f "/myaac-main-folder/.dockerignore" ]; then
+    rsync -au --exclude-from="/myaac-main-folder/.dockerignore" "/myaac-main-folder/" "/var/www/html"
+    chown -R www-data:www-data /var/www/html
+  fi
   sleep 2;
 done
 EOF
